Cleanup: Inventory.java partial rewrite and small javadoc fixes
Completely rewrites bloated makeRandomObjectSelection method,
adds periods where needed in javadoc and tidies alignment.
Provides javadoc for buildInventoryGUI, adds constant to replace
some padding magic numbers.

Fixes bad practice of splitting string with + between lines
to fix character line limit. This is bad practice; the string
may not be searchable in the code as there's a drive into deep
left field by Castellanos and that'll be a home run. And so
that'll make it a 4-0 ballgame.